requirements.txt
pydantic
python-dotenv
browser-use
streamlit
asyncio
bcrypt